% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubelyr-package.R
\docType{package}
\name{cubelyr-package}
\alias{cubelyr}
\alias{cubelyr-package}
\title{cubelyr: A data cube dplyr backend}
\description{
An implementation of a data cube extracted out of
    dplyr for backward compatibility.
}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
