% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubelyr-package.R
\docType{package}
\name{cubelyr-package}
\alias{cubelyr}
\alias{cubelyr-package}
\title{cubelyr: A Data Cube 'dplyr' Backend}
\description{
An implementation of a data cube extracted out of 'dplyr' for backward compatibility.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hadley/cubelyr}
  \item Report bugs at \url{https://github.com/hadley/cubelyr/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
